Fri, Jun 10, 2022, 7:11 PM +00:00

+++ home/coins.c
@@ -18,7 +18,7 @@ int main(void)
 
     t = t * 100;
         printf("test = %.2f\n",t);
-    int trocoInt = (int)t;
+    int trocoInt = t;
         printf("test = %i\n",trocoInt);
 
     if (trocoInt >= 25)
